% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegaspec.R
\name{as_character}
\alias{as_character}
\title{Convert vegaspec to character}
\usage{
as_character(spec, pretty = TRUE)
}
\arguments{
\item{spec}{object to be coerced to Vega/Vega-Lite specification}

\item{pretty}{\code{logical} indicates to use pretty (vs. minified) formatting}
}
\value{
\code{character}
}
\description{
Convert vegaspec to character
}
\examples{
  as_character(spec_mtcars)
}
