% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-vegaspec.R
\name{as_json}
\alias{as_json}
\title{Write out a vegaspec as JSON}
\usage{
as_json(spec, pretty = TRUE)
}
\arguments{
\item{spec}{object to be coerced to Vega/Vega-Lite specification}

\item{pretty}{\code{logical} indicates to use pretty (vs. minified) formatting}
}
\value{
\code{vegaspec} object
}
\description{
Write out a vegaspec as JSON
}
