% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegaspec.R
\name{as_list}
\alias{as_list}
\title{Convert vegaspec to list}
\usage{
as_list(spec)
}
\arguments{
\item{spec}{object to be coerced to Vega/Vega-Lite specification}
}
\value{
\code{jsonlite::json} object
}
\description{
Convert vegaspec to list
}
\examples{
  as_list(spec_mtcars)
}
