% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegaspec.R
\name{as_vegaspec}
\alias{as_vegaspec}
\title{Coerce to a Vega/Vega-Lite specification}
\usage{
as_vegaspec(spec, validate_spec = TRUE, convert_data = TRUE,
  consolidate_data = TRUE)
}
\arguments{
\item{spec}{object to be coerced to Vega/Vega-Lite specification}

\item{validate_spec}{\code{logical} indicates to validate_spec the specification}

\item{convert_data}{\code{logical} indicates to convert any \code{data.frame} to JSON}

\item{consolidate_data}{\code{logical} indicates to consolidate data into
top-level \code{datasets}}
}
\description{
Coerce to a Vega/Vega-Lite specification
}
