% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-package.R
\name{assert_packages}
\alias{assert_packages}
\title{Assert that packages are loaded}
\usage{
assert_packages(packages)
}
\arguments{
\item{packages}{\code{character} vector of package names to check}
}
\value{
\code{logical} indicating success
}
\description{
This function can be useful in writing package functions that use
functions from packages that you "suggest". It asserts that these packages
are available, and throws an informative error for those packages
that are not.
}
\examples{
\dontrun{
  assert_packages(c("base", "utils"))
}
}
\seealso{
\href{http://r-pkgs.had.co.nz/description.html#dependencies}{R Packages book}
}
\keyword{internal}
