% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autosize.R
\name{autosize}
\alias{autosize}
\title{Autosize a vegaspec}
\usage{
autosize(spec, width = NULL, height = NULL)
}
\arguments{
\item{spec}{object to be coerced to Vega/Vega-Lite specification}

\item{width, }{\code{integer} total width (pixels)}

\item{height, }{\code{integer} total height (pixels)}
}
\value{
object with S3 class \code{vegaspec}
}
\description{
Autosize a vegaspec
}
\examples{
  autosize(spec_mtcars, width = 350, height = 350)
}
