% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-block.R
\name{block_build_directory}
\alias{block_build_directory}
\title{Build a block directory}
\usage{
block_build_directory(path, spec, embed = vega_embed(),
  versions = vega_versions(major = FALSE), .block = block_config(),
  readme = NULL, use_thumbnail = TRUE, use_preview = TRUE)
}
\arguments{
\item{path}{\code{character}  path to an empty directory}

\item{spec}{object to be coerced to Vega/Vega-Lite specification}

\item{embed}{\code{vega_embed} object to specify embedding options -
the default is an empty call to \code{\link[=vega_embed]{vega_embed()}},
which will result in a canvas-rendering and action-links included for
exporting, viewing the Vega-Lite source, and opening the Vega editor.}

\item{versions}{named \code{list} of \code{character}:
names refer to JavaScript libraries \code{vega}, \code{vega_lite}, \code{vega_embed},
values are the tags for the versions to use in the block - use the
helper function \code{\link[=vega_versions]{vega_versions()}} with \code{major = TRUE} to use the current
major versions rather than the versions supported in this package}

\item{.block}{\code{character}, YAML text for the \code{.block} file -
use the helper function \code{\link[=block_config]{block_config()}} to specify block-configuration}

\item{readme}{\code{character}, single line, path to a markdown file;
multiple lines, markdown text}

\item{use_thumbnail}{\code{logical}, indicates to include a thumbnail image
(not yet implemented)}

\item{use_preview}{\code{logical}, indicates to include a preview image
(not yet implemented)}
}
\value{
\code{character} path, called for side effects
}
\description{
Build a block directory
}
\keyword{internal}
