% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-block.R
\name{block_config}
\alias{block_config}
\title{Specify the block configuration}
\usage{
block_config(license = "mit", height = 500, scrolling = TRUE,
  border = TRUE)
}
\arguments{
\item{license}{\code{character} specifies the license}

\item{height}{\code{numeric} height of the block (pixels)}

\item{scrolling}{\code{logical} indicates to use scrolling bars}

\item{border}{\code{logical} indicates to put a border on the \code{iframe}}
}
\value{
\code{character} yaml string for \code{.block} file
}
\description{
Specify the block configuration
}
\seealso{
\href{https://bl.ocks.org/-/about}{Blocks documentation}
}
