% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block.R
\name{block_create}
\alias{block_create}
\alias{block_create_gistid}
\title{Create gist to use as block}
\usage{
block_create(spec, embed = vega_embed(), .block = vw_block_config(),
  version = vega_version(major = FALSE), description = "", readme = NULL,
  use_thumbnail = TRUE, use_preview = TRUE, git_method = c("ssh",
  "https"), endpoint = NULL, env_pat = NULL, block_host = NULL,
  quiet = FALSE, browse = TRUE)

block_create_gistid(spec, embed = vega_embed(), .block = vw_block_config(),
  version = vega_version(major = FALSE), description = "", readme = NULL,
  use_thumbnail = TRUE, use_preview = TRUE, git_method = c("ssh",
  "https"), endpoint = NULL, env_pat = NULL, block_host = NULL,
  quiet = FALSE, browse = TRUE)
}
\arguments{
\item{spec}{object to be coerced to Vega/Vega-Lite specification}

\item{embed}{\code{vega_embed} object to specify embedding options -
the default is an empty call to \code{\link[=vega_embed]{vega_embed()}},
which will result in a canvas-rendering and action-links included for
exporting, viewing the Vega-Lite source, and opening the Vega editor.}

\item{.block}{\code{character}, YAML text for the \code{.block} file -
use the helper function \code{\link[=vw_block_config]{vw_block_config()}} to specify block-configuration}

\item{version}{named \code{list} of \code{character}:
names refer to JavaScript libraries \code{vega}, \code{vega_lite}, \code{vega_embed},
values are the tags for the versions to use in the block - use the
helper function \code{\link[=vega_version]{vega_version()}} with \code{major = TRUE} to use the current
major versions rather than the versions supported in this package}

\item{description}{\code{character}, description for the gist}

\item{readme}{\code{character}, single line, path to a markdown file;
multiple lines, markdown text}

\item{use_thumbnail}{\code{logical}, indicates to include a thumbnail image}

\item{use_preview}{\code{logical}, indicates to include a preview image}

\item{git_method}{\code{character}, use \code{"ssh"} or \code{"https"}}

\item{endpoint}{\code{character}, base endpoint for GitHub API, defaults to
\code{"https://api.github.com"}; useful to specify with GitHub Enterprise,
e.g. \code{"https://github.acme.com/api/v3"}
(not yet implemented)}

\item{env_pat}{\code{character}, name of environment variable that contains
a GitHub PAT (Personal Access Token), defaults to \code{"GITHUB_PAT"};
Useful to specify with GitHub Enterprise, e.g. \code{"GITHUB_ACME_PAT"}
(not yet implemented)}

\item{block_host}{\code{character}, hostname of blocks server,
defaults to \code{bl.ocks.org}}

\item{quiet}{\code{logical}, indicates to supress messages}

\item{browse}{\code{logical}, indicates to open browser to blocks web-page
when complete}
}
\value{
Called for side effects
\describe{
\item{\code{create_block()}}{returns an invisible copy of \code{spec}}
\item{\code{create_block_gistid()}}{returns a list with elements:
\describe{
\item{\code{endpoint}}{\code{character}, git API endpiont}
\item{\code{id}}{\code{character}, gist id}
}
}
}
}
\description{
These functions do the same thing: create a gist; they differ only in what
they return. \code{block_create()} returns a copy of \code{spec} so that it can be
used in a pipe; \code{block_create_gistid()} returns a list of information
about the newly-created gist.
}
\seealso{
\href{https://bl.ocks.org/-/about}{Blocks documentation},
\code{\link[gistr:gist_create_git]{gistr::gist_create_git()}}
}
