% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block.R
\name{block_retrieve}
\alias{block_retrieve}
\title{Retrieve spec from a block}
\usage{
block_retrieve(id, file = NULL, endpoint = NULL, env_pat = NULL)
}
\arguments{
\item{id}{\code{character} block id}

\item{file}{\code{character} filename within block,
\code{NULL} will retrieve first JSON file}

\item{endpoint}{\code{character}, base endpoint for GitHub API, defaults to
\code{"https://api.github.com"}; useful to specify with GitHub Enterprise,
e.g. \code{"https://github.acme.com/api/v3"}
(not yet implemented)}

\item{env_pat}{\code{character}, name of environment variable that contains
a GitHub PAT (Personal Access Token), defaults to \code{"GITHUB_PAT"};
Useful to specify with GitHub Enterprise, e.g. \code{"GITHUB_ACME_PAT"}
(not yet implemented)}
}
\value{
\code{vegaspec}
}
\description{
Need a github PAT in envvar \code{GITHUB_PAT}
}
