% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-list.R
\name{combine_lists}
\alias{combine_lists}
\title{Combine lists}
\usage{
combine_lists(.x, .y)
}
\arguments{
\item{.x}{\code{list()} to be combined}

\item{.y}{\code{list()} to be combined}
}
\value{
\code{list()}
}
\description{
Combine lists
}
