% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_category}
\alias{data_category}
\title{Example dataset: Categorical data}
\format{A data frame with ten observations of two variables
\describe{
\item{category}{\code{character}, representative of a nominal variable}
\item{number}{\code{double}, representative of a quantitative variable}
}}
\usage{
data_category
}
\description{
This is a toy dataset; the numbers are generated randomly.
}
\keyword{datasets}
