% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_seattle_daily}
\alias{data_seattle_daily}
\title{Example dataset: Seattle daily weather}
\format{A data frame with 1461 observations of six variables
\describe{
\item{date}{\code{Date}, date of the observation}
\item{precipitation}{\code{double}, amount of precipitation (mm)}
\item{temp_max}{\code{double}, maximum temperature (°C)}
\item{temp_min}{\code{double}, minimum temperature (°C)}
\item{wind}{\code{double}, average wind-speed (m/s)}
\item{weather}{\code{character}, description of weather}
}}
\source{
\url{https://vega.github.io/vega-datasets/data/seattle-weather.csv}
}
\usage{
data_seattle_daily
}
\description{
This dataset contains daily weather-observations from Seattle for the
years 2012-2015, inclusive.
}
\keyword{datasets}
