% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_seattle_hourly}
\alias{data_seattle_hourly}
\title{Example dataset: Seattle hourly temperatures}
\format{
A data frame with 8759 observations of two variables
\describe{
\item{date}{\code{POSIXct}, instant of the observation, uses \code{"America/Los_Angeles"}}
\item{temp}{\code{double}, temperature (°C)}
}
}
\source{
\url{https://vega.github.io/vega-datasets/data/seattle-weather-hourly-normals.csv}
}
\usage{
data_seattle_hourly
}
\description{
This dataset contains hourly temperature observations from Seattle for the
year 2010.
}
\keyword{datasets}
