% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js-handler-internal.R
\name{.vw_handler_body}
\alias{.vw_handler_body}
\title{Constructor for internal S3 class}
\usage{
.vw_handler_body(text, params = NULL)
}
\arguments{
\item{text}{\code{character}, text of the function body}

\item{params}{\code{character}, vector of names of required parameters}
}
\value{
S3 object with class \code{.vw_handler_body}
}
\description{
This S3 class is used to define handler-function bodies.
}
\seealso{
.vw_handler_def
}
\keyword{internal}
