% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vega-versions.R
\name{get_vega_versions}
\alias{get_vega_versions}
\title{Get versions of Vega JS libraries}
\usage{
get_vega_versions(vega_lite_version)
}
\arguments{
\item{vega_lite_version}{\code{character} version of Vega-Lite, e.g. \code{"2.5.0"}}
}
\value{
\code{list} with elements \code{vega_lite}, \code{vega}, \code{vega_embed}
}
\description{
This is an internal function for updating this package.
}
\examples{
\dontrun{
  vega_versions(vega_lite_version = "2.5.0")
}

}
\keyword{internal}
