% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/javascript.R
\name{glue_js}
\alias{glue_js}
\title{Interpolate into a JavaScript string}
\usage{
glue_js(..., .open = "${", .envir = parent.frame())
}
\arguments{
\item{...}{character vectors as the JavaScript source code (all arguments
will be pasted into one character string)}

\item{.open}{\code{character}, opening delimeter used by \code{\link[glue:glue]{glue::glue()}}}

\item{.envir}{\code{environment}, tells \code{\link[glue:glue]{glue::glue()}} where to find
the variables to be interpolated}
}
\value{
\code{\link[htmlwidgets:JS]{htmlwidgets::JS()}} object, type of charatcter vector
}
\description{
Uses (ES6) JavaScript notation to interpolate (R) variables into a string.
}
\details{
This is a wrapper to both \code{\link[htmlwidgets:JS]{htmlwidgets::JS()}} and \code{\link[glue:glue]{glue::glue()}}.
}
\examples{
  x <- 123
  glue_js("function(){return(${x});}")
}
