% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{knit_print.vegaspec}
\alias{knit_print.vegaspec}
\title{Knit-print method}
\usage{
knit_print.vegaspec(spec, ..., options = NULL)
}
\arguments{
\item{spec}{object to be coerced to Vega/Vega-Lite specification}

\item{...}{other arguments}

\item{options}{\code{list}, knitr options}
}
\description{
The only effective sizing options are \code{vega.width} and \code{vega.height};
these are passed to \code{\link[=vegawidget]{vegawidget()}} as \code{width} and \code{height},
respectively.
}
\details{
Embedding options can be passed by setting the option \code{vega.embed} using
\code{\link[=vega_embed]{vega_embed()}}.

\code{vega.width} and \code{vega.height} are
coerced to numeric, so it is ineffective to specify a percentage.
}
