% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-list.R
\name{pluck_all}
\alias{pluck_all}
\title{Pluck and combine recursively}
\usage{
pluck_all(.x, ..., .predicate = NULL, .equivalence = NULL)
}
\arguments{
\item{.x}{A vector or environment}

\item{...}{A list of accessors for indexing into the object. Can be
an integer position, a string name, or an accessor function
(except for the assignment variants which only support names and
positions). If the object being indexed is an S4 object,
accessing it by name will return the corresponding slot.

These dots support \link[rlang:list2]{tidy dots} features. In
particular, if your accessors are stored in a list, you can
splice that in with \code{!!!}.}

\item{.predicate}{\code{function} that takes a single argument, returns
a logical value indicating if this is the "thing" you are
looking for.}

\item{.equivalence}{\code{function} that takes two arguments, returns a
logical value indicating if these these things are equivalent.}
}
\value{
\code{list()}
}
\description{
Pluck and combine recursively
}
