% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_image.R
\name{png_bin}
\alias{png_bin}
\title{Translate PNG from dataURI string to binary}
\usage{
png_bin(png)
}
\arguments{
\item{png}{\code{character}, dataURI string describing PNG}
}
\value{
\code{raw} PNG
}
\description{
Translate PNG from dataURI string to binary
}
\keyword{internal}
