% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegawidget.R
\name{renderVegawidget}
\alias{renderVegawidget}
\title{Render shiny-output for vegawidget}
\usage{
renderVegawidget(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{An expression that generates an HTML widget (or a
\href{https://rstudio.github.io/promises/}{promise} of an HTML widget).}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Render shiny-output for vegawidget
}
