% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-setters.R
\name{shiny-setters}
\alias{shiny-setters}
\alias{vw_shiny_set_signal}
\alias{vw_shiny_set_data}
\title{Set information in a Vega chart}
\usage{
vw_shiny_set_signal(outputId, name, value, run = TRUE)

vw_shiny_set_data(outputId, name, value, use_cache = FALSE, run = TRUE)
}
\arguments{
\item{outputId}{\code{character}, shiny \code{outputId} for the vegawidget}

\item{name}{\code{character}, name of the signal or dataset being set,
as defined in the vegaspec}

\item{value}{reactive expression, e.g. \code{input$slider} or \code{dataset()},
that returns the value to which to set the signal or dataset}

\item{run}{\code{logical} indicates if the chart is to be run immediately}

\item{use_cache}{\code{logical}, for setting data, indicates to
to send Vega only the \emph{changes} in the dataset, rather
than making a hard reset of the dataset}
}
\value{
\code{\link[shiny:observeEvent]{shiny::observeEvent()}} that responds to changes in \code{expr}
}
\description{
There are two ways to change a Vega chart using shiny; by setting
a signal or by setting a dataset. In either case, you will need to
have named your signal or your dataset in the vegaspec.
}
\details{
These functions act as observers; use them as you would use
\code{\link[shiny:observe]{shiny::observe()}} or \code{\link[shiny:observeEvent]{shiny::observeEvent()}}.
}
