% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spec_mtcars}
\alias{spec_mtcars}
\title{Vega-Lite specification}
\format{A \code{vegaspec}, thin wrapper around list}
\usage{
spec_mtcars
}
\description{
A Vega-Lite specification to create a scatterplot for \code{mtcars}.
}
\keyword{datasets}
