% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{to_svg}
\alias{to_svg}
\title{to_svg}
\usage{
to_svg(spec, scale = 1, embed = NULL, width = NULL, height = NULL, ...)
}
\arguments{
\item{spec}{object to be coerced to Vega/Vega-Lite specification}

\item{scale}{scaleFactor}

\item{embed}{\code{vega_embed} object to specify embedding options -
the default is an empty call to \code{\link[=vega_embed]{vega_embed()}},
which will result in a canvas-rendering and action-links included for
exporting, viewing the Vega-Lite source, and opening the Vega editor.}

\item{width}{\code{integer}, if specified, the total rendered width (in pixels)
of the chart - valid only for single-view charts and layered charts;
the default is to use the width in the chart specification}

\item{height}{\code{integer}, if specified, the total rendered height (in pixels)
of the chart - valid only for single-view charts and layered charts;
the default is to use the height in the chart specification}

\item{...}{other arguments passed to \code{\link[htmlwidgets:createWidget]{htmlwidgets::createWidget()}}}
}
\description{
Turn a spec into an svg
}
