% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifiers.R
\name{vega-data-handlers}
\alias{vega-data-handlers}
\alias{vw_change_data}
\title{Dynamically modify data for a vegawidget with shiny}
\usage{
vw_change_data(input_id, name, newdata)
}
\arguments{
\item{input_id}{the div element the widget is bound to}

\item{name}{the named data set in the spec to change}

\item{newdata}{a data.frame to pass to the visualization}
}
\description{
Dynamically modify data for a vegawidget with shiny
}
\details{
Use inside a shiny application with shiny::observe({})
}
