% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifiers.R
\name{vega-data-loaders}
\alias{vega-data-loaders}
\alias{vw_load_data}
\title{Propagate a widget with data}
\usage{
vw_load_data(x, name, newdata)
}
\arguments{
\item{x}{a vegawidget object}

\item{name}{the named data set to import to}

\item{newdata}{a data.frame to pass to the visualization on render}
}
\description{
Propagate a widget with data
}
\details{
If no name field(s) has been specified in the data element
of the vega spec as defined by x, this function will throw an error.
}
