% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-getters.R
\name{vega-listeners}
\alias{vega-listeners}
\alias{vw_add_event_listener}
\alias{vw_add_signal_listener}
\title{Event and Signal Listeners}
\usage{
vw_add_event_listener(x, event, handler = "shiny")

vw_add_signal_listener(x, signal, handler = "shiny")
}
\arguments{
\item{x}{a vegawidget object}

\item{event}{name of an event, e.g. 'click'}

\item{handler}{either "shiny" (default) to add a listener that will expose
the relevant data to shiny via {output_id}\emph{{event} or {output_id}}{signal}}

\item{signal}{name of a signal specified in the vega spec (or created by
vega-lite when compiling into vega)}
}
\description{
Event and Signal Listeners
}
