% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vega-version.R
\name{vega_version}
\alias{vega_version}
\title{Get version-tags of supported Vega/Vega-Lite libraries}
\usage{
vega_version(major = FALSE)
}
\arguments{
\item{major}{\code{logical} return major version-tags rather than the
tags for the specific versions supported by this package}
}
\value{
\code{list} with \code{character} elements
named \code{vega_lite}, \code{vega}, \code{vega_embed}
}
\description{
Get version-tags of supported Vega/Vega-Lite libraries
}
\examples{
  vega_version()
  vega_version(major = TRUE)
}
