% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vega-versions.R
\name{vega_versions}
\alias{vega_versions}
\title{Get version-tags of supported Vega/Vega-Lite libraries}
\usage{
vega_versions(major = FALSE)
}
\arguments{
\item{major}{\code{logical} return major version-tags rather than the
tags for the specific versions supported by this package}
}
\value{
\code{list} with \code{character} elements
named \code{vega_lite}, \code{vega}, \code{vega_embed}
}
\description{
Get version-tags of supported Vega/Vega-Lite libraries
}
\examples{
  vega_versions()
  vega_versions(major = TRUE)
}
