% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-vegaspec.R
\name{vegaspec_data_to_values}
\alias{vegaspec_data_to_values}
\title{Fix list for vegaspec}
\usage{
vegaspec_data_to_values(spec)
}
\arguments{
\item{spec}{\code{list} for a Vega/Vega-Lite spec}
}
\value{
\code{list}
}
\description{
It is shorthand to specify data as a data-frame; however we need to
capture these "shorthands" in the spec and replace it with
a list with a single element named \code{values}.
}
\keyword{internal}
