% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_type.R
\name{vegaspec_type}
\alias{vegaspec_type}
\title{Determine library and version of a vegaspec}
\usage{
vegaspec_type(spec)
}
\arguments{
\item{spec}{object to be coerced to Vega/Vega-Lite specification}
}
\value{
named \code{list}
}
\description{
Examines the \code{$schema} element of a vegaspec.
}
\details{
Returns a list with two elements:

\describe{
\item{\code{library}}{\code{character}, either \code{"vega"} or \code{"vegalite"}}
\item{\code{version}}{\code{character}, version tag}
}
}
\examples{
  vegaspec_type(spec_mtcars)
}
