% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegawidget.R
\name{vegawidget}
\alias{vegawidget}
\title{Create a Vega/Vega-Lite htmlwidget}
\usage{
vegawidget(spec, embed = NULL, width = NULL, height = NULL,
  validate = TRUE, consolidate = TRUE, ...)
}
\arguments{
\item{spec}{chart_specification}

\item{embed}{\code{vega_embed} object to specify embedding options -
the default is an empty call to \code{\link[=vega_embed]{vega_embed()}},
which will result in a canvas-rendering and action-links included for
exporting, viewing the Vega-Lite source, and opening the Vega editor.}

\item{width}{\code{integer}, if specified, the total rendered width (in pixels)
of the chart - valid only for single-view charts and layered charts;
the default is to use the width in the chart specification}

\item{height}{\code{integer}, if specified, the total rendered height (in pixels)
of the chart - valid only for single-view charts and layered charts;
the default is to use the height in the chart specification}

\item{validate}{\code{logical} indicates to validate the specification}

\item{consolidate}{\code{logical} indicates to consolidate data into
top-level \code{datasets}}

\item{...}{other arguments}
}
\description{
Renders a chart as an htmlwidget.
}
\details{
This function is called \code{vegawidget()} is because it returns an htmlwidget
that uses Vega-Lite and Vega JavaScript libraries.

To specify embedding-options, use the \code{\link[=vega_embed]{vega_embed()}} function with the
\code{embed} argument. Its most-important options are:
\itemize{
\item \code{renderer}, to specify \code{"canvas"} (default) or \code{"svg"}
\item \code{actions},  to specify action-links
for \code{export}, \code{source}, \code{compiled}, and \code{editor}
}

If \code{actions} is \code{TRUE} (or \code{FALSE}), all the links
are shown (or not). Use a named list to be more specific, see
\code{\link[=vega_embed]{vega_embed()}} and \code{\link[=only_actions]{only_actions()}}.

If \code{embed} is \code{NULL}, \code{vegawidget()} sets \code{embed} to the value of
\code{getOption("altair.embed_options")}. Then, if this option is \code{NULL}, the
\href{https://github.com/vega/vega-embed#api-reference}{vega-embed}
defaults are used.

The arguments \code{width} and \code{height} are used to override the width and height
determined using the \code{chart} specification. However, there are some
important provisions:
\itemize{
\item Specifying \code{width} and \code{height} in \code{vegawidget()} is
\href{https://vega.github.io/vega-lite/docs/size.html#limitations}{effective only for single-view charts and layered charts}.
It will not work for contatenated, faceted, or repeated charts.
\item In the chart specification, the default interpretation of width and height
is to describe the dimensions of the
\strong{plotting rectangle}, not including the space used by the axes, labels,
etc. When \code{width} and \code{height} are specified using \code{vegawidget()},
the meanings change to describe the dimensions of the \strong{entire} rendered chart,
including axes, labels, etc.
\item Keep in mind that the action-links are not a part of the rendered chart,
so you may have to account for them yourself. You might expect
the height of the action-links to be 15-20 pixels.
}
}
