% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegaspec.R
\name{vw_as_json}
\alias{vw_as_json}
\title{Coerce vegaspec to JSON}
\usage{
vw_as_json(spec, pretty = TRUE)
}
\arguments{
\item{spec}{object to be coerced to Vega/Vega-Lite specification}

\item{pretty}{\code{logical} indicates to use pretty (vs. minified) formatting}
}
\value{
\code{jsonlite::json} object
}
\description{
Coerce vegaspec to JSON
}
\examples{
  vw_as_json(vw_ex_mtcars)
}
