% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view-api.R
\name{vw_bind_ui}
\alias{vw_bind_ui}
\title{Bind Shiny UI Element to Vega Signal}
\usage{
vw_bind_ui(output_id, input_id, signal_name, input_transformer = NULL)
}
\arguments{
\item{output_id}{Name of output element}

\item{input_id}{Name of input element}

\item{signal_name}{name of signal}

\item{input_transformer}{optional function to transform input before passing
to signal}
}
\value{
shiny observer that will update signal upon changes to input
}
\description{
Function to bind a shiny UI element to a Vega signal
}
