% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-block.R
\name{vw_block_build_directory}
\alias{vw_block_build_directory}
\title{Build block-directory}
\usage{
vw_block_build_directory(path, spec, embed = vega_embed(),
  version = vega_version(major = FALSE), .block = vw_block_config(),
  readme = NULL, use_thumbnail = TRUE, use_preview = TRUE)
}
\arguments{
\item{path}{\code{character}  path to an empty directory}

\item{spec}{object to be coerced to \code{vegaspec}, a Vega/Vega-Lite specification}

\item{embed}{\code{list} to specify
\href{https://github.com/vega/vega-embed#options}{vega-embed} options,
see \strong{Details} on how this is set if \code{NULL}.}

\item{version}{named \code{list} of \code{character}:
names refer to JavaScript libraries \code{vega}, \code{vega_lite}, \code{vega_embed},
values are the tags for the versions to use in the block - use the
helper function \code{\link[=vega_version]{vega_version()}} with \code{major = TRUE} to use the current
major versions rather than the versions supported in this package}

\item{.block}{\code{character}, YAML text for the \code{.block} file -
use the helper function \code{\link[=vw_block_config]{vw_block_config()}} to specify block-configuration}

\item{readme}{\code{character}, single line, path to a markdown file;
multiple lines, markdown text}

\item{use_thumbnail}{\code{logical}, indicates to include a thumbnail image}

\item{use_preview}{\code{logical}, indicates to include a preview image}
}
\value{
\code{character} path, called for side effects
}
\description{
Build block-directory
}
\keyword{internal}
