% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block.R
\name{vw_create_block}
\alias{vw_create_block}
\alias{vw_create_block_gistid}
\title{Create gist to use as block}
\usage{
vw_create_block(spec, embed = vega_embed(), .block = vw_block_config(),
  version = vega_version(major = FALSE), description = NULL,
  readme = NULL, use_thumbnail = TRUE, use_preview = TRUE,
  git_method = c("ssh", "https"), endpoint = NULL, env_pat = NULL,
  block_host = NULL, quiet = FALSE, browse = TRUE)

vw_create_block_gistid(spec, embed = vega_embed(),
  .block = vw_block_config(), version = vega_version(major = FALSE),
  description = NULL, readme = NULL, use_thumbnail = TRUE,
  use_preview = TRUE, git_method = c("ssh", "https"),
  endpoint = NULL, env_pat = NULL, block_host = NULL,
  quiet = FALSE, browse = TRUE)
}
\arguments{
\item{spec}{object to be coerced to \code{vegaspec}, a Vega/Vega-Lite specification}

\item{embed}{\code{list} to specify
\href{https://github.com/vega/vega-embed#options}{vega-embed} options,
see \strong{Details} on how this is set if \code{NULL}.}

\item{.block}{\code{character}, YAML text for the \code{.block} file -
use the helper function \code{\link[=vw_block_config]{vw_block_config()}} to specify block-configuration}

\item{version}{named \code{list} of \code{character}:
names refer to JavaScript libraries \code{vega}, \code{vega_lite}, \code{vega_embed},
values are the tags for the versions to use in the block - use the
helper function \code{\link[=vega_version]{vega_version()}} with \code{major = TRUE} to use the current
major versions rather than the versions supported in this package}

\item{description}{\code{character}, description for the gist - if \code{NULL},
looks for a description field in \code{spec}.}

\item{readme}{\code{character}, single line, path to a markdown file;
multiple lines, markdown text}

\item{use_thumbnail}{\code{logical}, indicates to include a thumbnail image}

\item{use_preview}{\code{logical}, indicates to include a preview image}

\item{git_method}{\code{character}, use \code{"ssh"} or \code{"https"}}

\item{endpoint}{\code{character}, base endpoint for GitHub API, defaults to
\code{"https://api.github.com"}; useful to specify with GitHub Enterprise,
e.g. \code{"https://github.acme.com/api/v3"}
(not yet implemented)}

\item{env_pat}{\code{character}, name of environment variable that contains
a GitHub PAT (Personal Access Token), defaults to \code{"GITHUB_PAT"};
Useful to specify with GitHub Enterprise, e.g. \code{"GITHUB_ACME_PAT"}
(not yet implemented)}

\item{block_host}{\code{character}, hostname of blocks server,
defaults to \code{bl.ocks.org}}

\item{quiet}{\code{logical}, indicates to supress messages}

\item{browse}{\code{logical}, indicates to open browser to blocks web-page
when complete}
}
\value{
Called for side effects
\describe{
\item{\code{create_block()}}{returns an invisible copy of \code{spec}}
\item{\code{create_block_gistid()}}{returns a list with elements:
\describe{
\item{\code{endpoint}}{\code{character}, git API endpiont}
\item{\code{id}}{\code{character}, gist id}
}
}
}
}
\description{
These functions require that \href{https://cran.r-project.org/package=fs}{fs}
and  \href{https://cran.r-project.org/package=gistr}{gistr} be installed.
If you want to include a thumbnail or preview image, you will need the
\href{https://cran.r-project.org/package=magick}{magick} package,
\href{https://cran.r-project.org/package=webdriver}{webdriver} package,
and PhantomJS installed.
}
\details{
These functions do the same thing: create a gist; they differ only in what
they return. \code{vw_create_block()} returns a copy of \code{spec} so that it can be
used in a pipe; \code{vw_create_block_gistid()} returns a list of information
about the newly-created gist.

In addition to having the \strong{gistr} package installed, you will need a
GitHub Personal Access Token (PAT) stored in an environment variable
called \code{GITHUB_PAT}. See \href{http://happygitwithr.com/github-pat.html}{Happy Git with R}
for more information on how to acquire and store a PAT.

The default versions of the Vega JavaScript libraries are this package's
supported versions. To use the major (current) versions instead, use
\code{version = vega_version(major = TRUE)}.
}
\examples{
\dontrun{
  # use supported versions
  vw_create_block(spec_mtcars)
  # use major versions
  vw_create_block(spec_mtcars, version = vega_version(major = TRUE))
  # return gist information
  vw_create_block_gistid(spec_mtcars)
}

}
\seealso{
\href{https://bl.ocks.org/-/about}{Blocks documentation},
\code{\link[=vw_retrieve_block]{vw_retrieve_block()}},
\code{\link[gistr:gist_create_git]{gistr::gist_create_git()}}, \code{\link[=vega_version]{vega_version()}}, \code{\link[=vw_block_config]{vw_block_config()}}
}
