% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vw_ex_mtcars}
\alias{vw_ex_mtcars}
\title{Example vegaspec: mtcars scatterplot}
\format{A \code{vegaspec}, see \code{\link[=as_vegaspec]{as_vegaspec()}}}
\usage{
vw_ex_mtcars
}
\description{
A Vega-Lite specification to create a scatterplot for \code{mtcars}.
}
\keyword{datasets}
