% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js-handler.R
\name{vw_handler_body_compose}
\alias{vw_handler_body_compose}
\alias{vw_handler_compose}
\title{Compose a JavaScript handler}
\usage{
vw_handler_body_compose(vw_handler, n_indent = 2L)

vw_handler_compose(vw_handler)
}
\arguments{
\item{vw_handler}{\code{vw_handler} created using \code{\link[=vw_handler_signal]{vw_handler_signal()}} or
\code{\link[=vw_handler_event]{vw_handler_event()}}}

\item{n_indent}{\code{integer}, number of spaces to indent the text of the body}
}
\value{
object with S3 class \code{JS_EVAL}, text for the function or
function-body
}
\description{
These functions are used to compose a \code{vw_handler} object into
text that will be interpreted either as a complete JavaScript
function, or a function-body.
}
\keyword{internal}
