% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view-api.R
\name{vw_observe}
\alias{vw_observe}
\title{a potentially-useful experiment}
\usage{
vw_observe(expr, output_id, name, use_cache = TRUE)
}
\arguments{
\item{expr}{reactive expression, i.e. \code{input$slider} or \code{dataset()}}

\item{output_id}{\code{character} outputId for the vegawidget}

\item{name}{\code{character} name of the signal defined in the
vegawidget spec}
}
\value{
\code{\link[shiny:observeEvent]{shiny::observeEvent()}} that reacts to changes in \code{input}
}
\description{
a potentially-useful experiment
}
