% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view-api.R
\name{vw_observe_signal}
\alias{vw_observe_signal}
\title{a potentially-useful experiment}
\usage{
vw_observe_signal(input, output_id, signal_name,
  input_transformer = NULL)
}
\arguments{
\item{input}{a reactive value}

\item{output_id}{\code{character} outputId for the vegawidget}

\item{signal_name}{\code{character} name of the signal defined in the
vegawidget spec}

\item{input_transformer}{\code{function} (optional) applied to the value of \code{input}
before passing to Vega}
}
\value{
\code{\link[shiny:observeEvent]{shiny::observeEvent()}} that reacts to changes in \code{input}
}
\description{
a potentially-useful experiment
}
