% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view-api.R
\name{vw_observe_to_signal}
\alias{vw_observe_to_signal}
\title{a potentially-useful experiment}
\usage{
vw_observe_to_signal(expr, output_id, signal_name)
}
\arguments{
\item{expr}{reactive expression, i.e. \code{input$slider} or \code{dataset()}}

\item{output_id}{\code{character} outputId for the vegawidget}

\item{signal_name}{\code{character} name of the signal defined in the
vegawidget spec}
}
\value{
\code{\link[shiny:observeEvent]{shiny::observeEvent()}} that reacts to changes in\code{expr}
}
\description{
a potentially-useful experiment
}
