% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-image.R
\name{vw_png_bin}
\alias{vw_png_bin}
\title{Coerce data-URI string to raw binary}
\usage{
vw_png_bin(png)
}
\arguments{
\item{png}{\code{character}, data-URI string describing PNG}
}
\value{
\code{raw} PNG
}
\description{
\code{\link[=vw_to_png]{vw_to_png()}} returns a data-URI string for the PNG image. If
you want this PNG image as \code{raw} binary data, use \code{vw_to_png()} followed
this function.
}
\examples{
\dontrun{
   spec_mtcars \%>\% vw_to_png() \%>\% vw_png_bin()
}
}
