% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename-datasets.R
\name{vw_rename_datasets}
\alias{vw_rename_datasets}
\title{Rename datasets in a vegaspec}
\usage{
vw_rename_datasets(spec)
}
\arguments{
\item{spec}{An object to be coerced to \code{vegaspec}, a Vega/Vega-Lite specification}
}
\value{
S3 object of class \code{vegaspec}
}
\description{
If a vegaspec has named datasets, it may be useful to rename them.
This function will return a vegaspec with datasets named \code{data_001},
\code{data_002}, and so on. It will go through the spec and replace the
references to the names. A future version of this function may give you
the more control over the names used.
}
