% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize_data.R
\name{vw_serialize_data}
\alias{vw_serialize_data}
\title{Serialize data-frame time-columns}
\usage{
vw_serialize_data(data, iso_dttm = FALSE, iso_date = TRUE)
}
\arguments{
\item{data}{\code{data.frame}, data to be serialized}

\item{iso_dttm}{\code{logical}, indicates if datetimes (\code{POSIXct}) are to be
formatted using ISO-8601}

\item{iso_date}{\code{logical}, indicates if dates (\code{Date}) are to be
formatted using ISO-8601}
}
\value{
object with the same type as \code{data}
}
\description{
In Vega, for now, there are only two time-zones available: the local
time-zone of the browser where the spec is rendered, and UTC. This differs
from R, where a time-zone attribute is available to \code{POSIXct} vectors.
Accordingly, when designing a vegaspec that uses time, you have to make some
some compromises. This function helps you to implement your compromise in
a principled way, as explained in the opinions below.
}
\seealso{
\href{https://vega.github.io/vega-lite/docs/timeunit.html#utc}{Vega-Lite Time Unit (UTC)}
}
