% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-demo.R
\name{vw_shiny_demo}
\alias{vw_shiny_demo}
\title{Run Shiny demonstration-apps}
\usage{
vw_shiny_demo(example = NULL, ...)
}
\arguments{
\item{example}{\code{character}, name of the example to run; if NULL (default),
prints out a list of available examples}

\item{...}{additional arguments passed to \code{\link[shiny:runApp]{shiny::runApp()}}}
}
\value{
invisible NULL, called for side-effects
}
\description{
Run Shiny demonstration-apps
}
\examples{
  vw_shiny_demo() # returns available examples

  # Run only in interactive R sessions
  if (interactive()) {
    vw_shiny_demo("data-set-get")
  }
}
