% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec-version.R
\name{vw_spec_version}
\alias{vw_spec_version}
\title{Determine vegaspec version}
\usage{
vw_spec_version(spec)
}
\arguments{
\item{spec}{object to be coerced to Vega/Vega-Lite specification}
}
\value{
\code{list} with elements \code{library}, \code{version}
}
\description{
Examines the \code{$schema} element of a vegaspec.
}
\details{
Returns a list with two elements:

\describe{
\item{\code{library}}{\code{character}, either \code{"vega"} or \code{"vega_lite"}}
\item{\code{version}}{\code{character}, version tag}
}
}
\examples{
  vw_spec_version(spec_mtcars)
}
