% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to-image.R
\name{vw_to_svg}
\alias{vw_to_svg}
\alias{vw_to_svg.default}
\alias{vw_to_svg.vegawidget}
\title{Convert to SVG string}
\usage{
vw_to_svg(...)

\method{vw_to_svg}{default}(spec, scale = 1, embed = NULL, width = NULL,
  height = NULL, ...)

\method{vw_to_svg}{vegawidget}(widget, scale = 1, ...)
}
\arguments{
\item{...}{other arguments passed to \code{\link[htmlwidgets:createWidget]{htmlwidgets::createWidget()}}}

\item{spec}{object to be coerced to Vega/Vega-Lite specification}

\item{scale}{scale factor for the image}

\item{embed}{\code{vega_embed} object to specify embedding options -
the default is an empty call to \code{\link[=vega_embed]{vega_embed()}},
which will result in a canvas-rendering and action-links included for
exporting, viewing the Vega-Lite source, and opening the Vega editor.}

\item{width}{\code{integer}, if specified, the total rendered width (in pixels)
of the chart - valid only for single-view charts and layered charts;
the default is to use the width in the chart specification}

\item{height}{\code{integer}, if specified, the total rendered height (in pixels)
of the chart - valid only for single-view charts and layered charts;
the default is to use the height in the chart specification}

\item{widget}{object created using \code{\link[=vegawidget]{vegawidget()}}}
}
\value{
\code{character} SVG string
}
\description{
Convert to SVG string
}
\examples{
\dontrun{
  vw_to_svg(vw_ex_mtcars)
  vw_to_svg(vegawidget(vw_ex_mtcars))
}
}
