% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to-vega.R
\name{vw_to_vega}
\alias{vw_to_vega}
\title{Convert to Vega specification}
\usage{
vw_to_vega(spec)
}
\arguments{
\item{spec}{object to be coerced to Vega/Vega-Lite specification}
}
\value{
a vegaspec for Vega
}
\description{
Convert to Vega specification
}
