% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-image.R
\name{vw_write_png}
\alias{vw_write_png}
\alias{vw_write_png.default}
\alias{vw_write_png.vegawidget}
\title{Write to PNG file}
\usage{
vw_write_png(...)

\method{vw_write_png}{default}(spec, path, scale = 1, embed = NULL,
  width = NULL, height = NULL, ...)

\method{vw_write_png}{vegawidget}(widget, path, scale = 1, ...)
}
\arguments{
\item{...}{other arguments passed to \code{\link[htmlwidgets:createWidget]{htmlwidgets::createWidget()}}}

\item{spec}{object to be coerced to Vega/Vega-Lite specification}

\item{path}{path to which to write image-file}

\item{scale}{scale factor for the image}

\item{embed}{\code{vega_embed} object to specify embedding options -
the default is an empty call to \code{\link[=vega_embed]{vega_embed()}},
which will result in a canvas-rendering and action-links included for
exporting, viewing the Vega-Lite source, and opening the Vega editor.}

\item{width}{\code{integer}, if specified, the total rendered width (in pixels)
of the chart - valid only for single-view charts and layered charts;
the default is to use the width in the chart specification}

\item{height}{\code{integer}, if specified, the total rendered height (in pixels)
of the chart - valid only for single-view charts and layered charts;
the default is to use the height in the chart specification}

\item{widget}{object created using \code{\link[=vegawidget]{vegawidget()}}}
}
\value{
copy of whatever was sent to it
}
\description{
Write to PNG file
}
\examples{
\dontrun{
  write_svg(vw_ex_mtcars)
  write_svg(vegawidget(vw_ex_mtcars))
}
}
