% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{write_png}
\alias{write_png}
\title{Write a spec to a png file}
\usage{
write_png(spec, path, embed = NULL, width = NULL, height = NULL, ...)
}
\arguments{
\item{spec}{object to be coerced to Vega/Vega-Lite specification}

\item{path}{\code{character} path to write to}

\item{embed}{\code{vega_embed} object to specify embedding options -
the default is an empty call to \code{\link[=vega_embed]{vega_embed()}},
which will result in a canvas-rendering and action-links included for
exporting, viewing the Vega-Lite source, and opening the Vega editor.}

\item{width}{\code{integer}, if specified, the total rendered width (in pixels)
of the chart - valid only for single-view charts and layered charts;
the default is to use the width in the chart specification}

\item{height}{\code{integer}, if specified, the total rendered height (in pixels)
of the chart - valid only for single-view charts and layered charts;
the default is to use the height in the chart specification}

\item{...}{other arguments passed to \code{\link[htmlwidgets:createWidget]{htmlwidgets::createWidget()}}}
}
\value{
invisible copy of \code{spec}
}
\description{
This function has the side-effect of writing out an image file,
it returns an invisible copy of \code{spec} so that it can be part of a pipe.
}
\examples{
\dontrun{
  write_png(spec_mtcars, "mtcars.png")
}


}
