% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegaspec.R
\name{as_vegaspec}
\alias{as_vegaspec}
\alias{as_vegaspec.default}
\alias{as_vegaspec.vegaspec}
\alias{as_vegaspec.list}
\alias{as_vegaspec.json}
\alias{as_vegaspec.character}
\title{Coerce to vegaspec}
\usage{
as_vegaspec(spec, ...)

\method{as_vegaspec}{default}(spec, ...)

\method{as_vegaspec}{vegaspec}(spec, ...)

\method{as_vegaspec}{list}(spec, ...)

\method{as_vegaspec}{json}(spec, ...)

\method{as_vegaspec}{character}(spec, ...)
}
\arguments{
\item{spec}{object to be coerced to \code{vegaspec}, a Vega/Vega-Lite specification}

\item{...}{other args (attempt to future-proof)}
}
\value{
S3 object of class \code{vegaspec}
}
\description{
Vega and Vega-Lite use JSON as the format for their specifications. Within R,
it seems natural to work with these specifications as lists. Accordingly,
a \code{vegaspec} is also a list. This family of functions is used to coerce lists,
JSON, and character strings to \code{vegaspec}.
}
\details{
The \code{character} method for this function will take:
\itemize{
\item{JSON string}
\item{A path to a local JSON file}
\item{A URL that contains a JSON file, requires \href{https://CRAN.R-project.org/package=httr}{httr} be installed}
}

For Vega and Vega-Lite, the translation between list and JSON is a little
bit particular. This function, \code{\link[=as_vegaspec]{as_vegaspec()}}, can be used to translate
from JSON; \code{\link[=vw_as_json]{vw_as_json()}} can be used to translate to JSON.

A given \code{vegaspec} will be particular to either Vega or Vega-Lite. You can use
the function \code{\link[=vw_spec_version]{vw_spec_version()}} to determine this. You can use
\code{\link[=vw_to_vega]{vw_to_vega()}} to translate a Vega-Lite spec to Vega.
}
\examples{
  \dontrun{
    as_vegaspec("https://vega.github.io/vega-lite/examples/specs/bar.vl.json")
  }
}
\seealso{
\href{https://vega.github.io/vega/}{Vega},
\href{https://vega.github.io/vega-lite/}{Vega-Lite},
\code{\link[=vw_as_json]{vw_as_json()}}, \code{\link[=vw_spec_version]{vw_spec_version()}}, \code{\link[=vw_to_vega]{vw_to_vega()}}
}
