% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegaspec.R
\name{as_vegaspec}
\alias{as_vegaspec}
\title{Coerce to a Vega/Vega-Lite specification}
\usage{
as_vegaspec(spec, validate = TRUE, consolidate = TRUE)
}
\arguments{
\item{spec}{object to be coerced to Vega/Vega-Lite specification}

\item{validate}{\code{logical} indicates to validate the specification}

\item{consolidate}{\code{logical} indicates to consolidate data into
top-level \code{datasets}}
}
\value{
S3 object of class \code{vegaspec}
}
\description{
Talk about how \code{vegaspec} is a thin wrapper to \code{list}. Implemented as JSON.
}
\details{
Talk about this is the chance to validate the spec and to consolidate the
datasets.
}
