% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autosize.R
\name{autosize}
\alias{autosize}
\title{Autosize vegaspec}
\usage{
autosize(spec, width = NULL, height = NULL)
}
\arguments{
\item{spec}{object to be coerced to Vega/Vega-Lite specification}

\item{width}{\code{integer}, if specified, the total rendered width (in pixels)
of the chart - valid only for single-view charts and layered charts;
the default is to use the width in the chart specification}

\item{height}{\code{integer}, if specified, the total rendered height (in pixels)
of the chart - valid only for single-view charts and layered charts;
the default is to use the height in the chart specification}
}
\value{
object with S3 class \code{vegaspec}
}
\description{
The arguments \code{width} and \code{height} can be used to override the width and height
of the provided \code{spec}.
}
\details{
There are some important limitations:
\itemize{
\item Specifying \code{width} and \code{height} is
\href{https://vega.github.io/vega-lite/docs/size.html#limitations}{effective only for single-view and layered specifications}.
It will not work for specifications with multiple views
(e.g. \code{hconcat}, \code{vconcat}, \code{facet}, \code{repeat}); this will issue a
warning that there will be no effect on the specification when rendered.
\item In the specification, the default interpretation of width and height
is to describe the dimensions of the
\strong{plotting rectangle}, not including the space used by the axes, labels,
etc. When \code{width} and \code{height} are specified,
the meanings change to describe the dimensions of the \strong{entire} rendered chart,
including axes, labels, etc.
}
}
\examples{
  autosize(spec_mtcars, width = 350, height = 350)
}
