% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block.R
\name{create_block}
\alias{create_block}
\title{Create a gist that can be used as a block}
\usage{
create_block(spec, embed = vega_embed(), description = NULL,
  block = block_yaml(), readme = NULL, use_thumbnail = TRUE,
  use_preview = TRUE, use_current = FALSE, git_method = c("ssh", "https"),
  host = NULL, auth_token = NULL)
}
\arguments{
\item{spec}{\code{json}, \code{list}, or \code{character} Vega-Lite or Vega specification}

\item{embed}{\code{vega_embed} object}

\item{description}{\code{character} description for the gist}

\item{block}{\code{character} YAML text for the \code{.block} file}

\item{readme}{\code{character} single line, path to a markdown file;
multiple lines, markdown text}

\item{use_thumbnail}{\code{logical} indicates to include a thumbnail image}

\item{use_preview}{\code{logical} indicates to include a preview image}

\item{use_current}{\code{logical} or named \code{list} of \code{logical},
for names of JavaScript libraries \code{vega}, \code{vega_lite}, \code{vega_embed}:
\itemize{
\item \code{TRUE}: use the current major version
\item \code{FALSE}: use the version use in the htmlwidget for this package
}}

\item{git_method}{\code{character} use \code{"ssh"} or \code{"https"}}

\item{host}{\code{character} host address,
defaults to \code{"gist.github.com"}}

\item{auth_token}{\code{character} GitHub PAT,
defaults to \code{Sys.getenv("GITHUB_PAT")}}
}
\value{
called for side effects
}
\description{
idea: function to validate the spec
}
\seealso{
\href{https://bl.ocks.org/-/about}{Blocks documentation},
\code{\link[gistr:gist_create_git]{gistr::gist_create_git()}}
}
