% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to-image.R, R/write-image.R
\name{image}
\alias{image}
\alias{vw_to_svg}
\alias{vw_to_bitmap}
\alias{vw_write_svg}
\alias{vw_write_png}
\title{Create or write image}
\usage{
vw_to_svg(spec, width = NULL, height = NULL, base_url = "",
  seed = sample(1e+08, size = 1))

vw_to_bitmap(spec, width = NULL, height = NULL, ...)

vw_write_svg(spec, path, ...)

vw_write_png(spec, path, width = NULL, height = NULL, dpi = NULL,
  ...)
}
\arguments{
\item{spec}{\code{vegaspec}, object that is or can be coerced to a \code{vegaspec},
e.g. a vegawidget object.}

\item{width, height}{\code{numeric}, output width and height in pixels (or NULL for
default)}

\item{base_url}{\code{character}, the base url for a data file. Useful for
specifying a local directory}

\item{seed}{\code{integer}, the random seed for a vega spec}

\item{...}{additional argument to pass to \code{vw_to_svg}}

\item{path}{\code{character}, local path to which to write file}

\item{dpi}{\code{numeric}, numeric vector of length 1 or 2 specifying the
resolution of the image in DPI (dots per length) for x and y
(in that order). See \code{\link[png]{writePNG}}}
}
\value{
\describe{
\item{\code{vw_to_bitmap()}}{\code{array}, bitmap array}
\item{\code{vw_to_svg()}}{\code{character}, SVG string}
\item{\code{vw_write_png()}}{invisible \code{vegaspec} or \code{vegawidget}}
\item{\code{vw_write_svg()}}{invisible \code{vegaspec} or \code{vegawidget}}
}
}
\description{
If \url{node} is installed, these functions can be used to create
or write images as PNG or SVG, using a \code{vegaspec} or \code{vegawidget}.
}
\details{
These functions can be called using (an object that can be coerced to)
a \code{vegaspec}.

For writing a \code{png}, the \code{rsvg} package is needed in addition to
\strong{node}.

The \strong{node} scripts used are adapted from the command line utilies in the
vega-cli package.
}
\examples{
\dontrun{
  # call any of these functions using either a vegaspec or a vegawidget
  vw_to_svg(vegawidget(spec_mtcars))
  vw_to_bitmap(spec_mtcars)
  vw_write_png(spec_mtcars, "temp.png")
  vw_write_svg(spec_mtcars, "temp.svg")

  # To specify the path to a local file, use base_url
  spec_precip <- '{
  "$schema": "https://vega.github.io/schema/vega-lite/v3.json",
  "data": {"url": "seattle-weather.csv"},
  "mark": "tick",
  "encoding": {
    "x": {"field": "precipitation", "type": "quantitative"}
  }
  }"'
  data_dir <- system.file("example-data/", package = "vegawidget")
  vw_write_png(spec_precip, "temp-local.png", base_url = data_dir)

}
}
\seealso{
\href{https://github.com/vega/vega-view#image-export}{vega-view library},
}
