% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-getters.R
\name{shiny-getters}
\alias{shiny-getters}
\alias{vw_shiny_get_signal}
\alias{vw_shiny_get_event}
\title{Get a signal, data, or value from an event using a reactive expression}
\usage{
vw_shiny_get_signal(outputId, name, handler = "value")

vw_shiny_get_event(outputId, name, handler = "datum")
}
\arguments{
\item{outputId}{\code{character}, shiny \code{outputId} for the vegawidget}

\item{name}{\code{character}, name of the signal or dataset being monitored}

\item{handler}{\code{character}}
}
\value{
\code{\link[shiny:reactive]{shiny::reactive()}} that returns the value returned by the
\code{handler}
}
\description{
Get a signal, data, or value from an event using a reactive expression
}
