% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spec_mtcars}
\alias{spec_mtcars}
\title{Example vegaspec: mtcars scatterplot}
\format{
S3 object of class \code{vegaspec}
}
\usage{
spec_mtcars
}
\description{
A Vega-Lite specification to create a scatterplot for \code{mtcars}.
}
\seealso{
\code{\link[=as_vegaspec]{as_vegaspec()}}
}
\keyword{datasets}
