% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifiers.R
\name{vega-data-handlers}
\alias{vega-data-handlers}
\alias{vw_insert_data}
\title{Modify data for a vegawidget}
\usage{
vw_insert_data(x, name, newdata)
}
\arguments{
\item{x}{a vegawidget object}

\item{name}{the named data set to import to}

\item{newdata}{a data.frame to pass to the visualization}
}
\description{
Modify data for a vegawidget
}
\details{
If no name field(s) has been specified in the data element
of the vega spec as defined by x, this function will throw an error.
}
