% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vega-versions.R
\name{vega_versions}
\alias{vega_versions}
\title{Get versions of supported Vega/Vega-Lite libraries}
\usage{
vega_versions(use_major = FALSE)
}
\arguments{
\item{use_major}{\code{logical} return major, rather than current, versions}
}
\value{
\code{list} with \code{character} elements
named \code{vega_lite}, \code{vega}, \code{vega_embed}
}
\description{
Get versions of supported Vega/Vega-Lite libraries
}
\examples{
  vega_versions()
  vega_versions(use_major = TRUE)
}
