% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js-handler-internal.R
\name{vw_handler}
\alias{vw_handler}
\title{Construct a vegawidget handler}
\usage{
vw_handler(args, body_value, body_effect)
}
\arguments{
\item{args}{\code{character}, vector of names for the arguments for the handler}

\item{body_value}{\code{character}, body of the \emph{value} part of the handler}

\item{body_effect}{\code{character}, body of the \emph{effect} part of the handler}
}
\value{
object with S3 class \code{vw_handler}
}
\description{
You will likely call one of the specific handler-constructors:
\code{\link[=vw_handler_signal]{vw_handler_signal()}}, \code{\link[=vw_handler_event]{vw_handler_event()}}, in conjunction with
\code{\link[=vw_handler_add_effect]{vw_handler_add_effect()}}.
}
\details{
The handler has three parts:
\itemize{
\item \code{args} arguments to the handler
\item \code{body_value} the body of a function of the \code{args}, returns a value
\item \code{body_effect} the body of a function of the value, \code{x}, performs a
side-effect
}
}
\seealso{
\code{\link[=vw_handler_signal]{vw_handler_signal()}}, \code{\link[=vw_handler_event]{vw_handler_event()}},
\code{\link[=vw_handler_add_effect]{vw_handler_add_effect()}}
}
\keyword{internal}
