% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-image.R
\name{vw_png_bin}
\alias{vw_png_bin}
\title{Coerce data-URI string to raw binary}
\usage{
vw_png_bin(png)
}
\arguments{
\item{png}{\code{character}, data-URI string describing PNG}
}
\value{
\code{raw} PNG
}
\description{
Coerce data-URI string to raw binary
}
\examples{
\dontrun{
  png_bin <-
    spec_mtcars \%>\%
    vw_to_png() \%>\%
    vw_png_bin()
}
}
