% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to-vega.R
\name{vw_to_vega}
\alias{vw_to_vega}
\title{Convert to Vega specification}
\usage{
vw_to_vega(spec)
}
\arguments{
\item{spec}{object to be coerced to \code{vegaspec}, a Vega/Vega-Lite specification}
}
\value{
S3 object of class \code{vegaspec_vega} and \code{vegaspec}
}
\description{
If the \href{https://CRAN.R-project.org/package=V8}{V8} package is installed,
use this function to translate a Vega-Lite specification to a
Vega specification.
}
\examples{
  vw_spec_version(spec_mtcars)
  vw_spec_version(vw_to_vega(spec_mtcars))
}
