% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to-vega.R
\name{vw_to_vega}
\alias{vw_to_vega}
\title{Convert to Vega specification}
\usage{
vw_to_vega(spec)
}
\arguments{
\item{spec}{object to be coerced to \code{vegaspec}, a Vega/Vega-Lite specification}
}
\value{
S3 object of class \code{vegaspec_vega} and \code{vegaspec}
}
\description{
If you have  \strong{\href{https://nodejs.org/en/}{nodejs}} installed,
you can use this function to compile a Vega-Lite specification
into a Vega specification.
}
\examples{
  vw_spec_version(spec_mtcars)
\dontrun{
  # requires nodejs to be installed
  vw_spec_version(vw_to_vega(spec_mtcars))
}
}
